# src/Services/swagger/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Pick_IAdmin.email-or-password_": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"FlattenMaps_T_": {
				"properties": {},
				"type": "object"
			},
			"IClient": {
				"$ref": "#/components/schemas/FlattenMaps_T_"
			},
			"IWorker": {
				"$ref": "#/components/schemas/FlattenMaps_T_"
			},
			"IPost": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					},
					"clientId": {
						"type": "string"
					},
					"adresse": {
						"type": "string"
					},
					"photo": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"comments": {
						"items": {},
						"type": "array"
					},
					"status": {
						"type": "string",
						"enum": [
							"pending",
							"reserved",
							"completed"
						]
					}
				},
				"required": [
					"createdAt",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IMaterialPost": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					},
					"workerId": {
						"type": "string"
					},
					"adresse": {
						"type": "string"
					},
					"photo": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					}
				},
				"required": [
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IWorker.email-or-password_": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Payload": {
				"$ref": "#/components/schemas/Pick_IWorker.email-or-password_"
			},
			"Pick_IClient.nom-or-prenom-or-password-or-adresse-or-email_": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"nom": {
						"type": "string"
					},
					"prenom": {
						"type": "string"
					},
					"adresse": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"IComment": {
				"properties": {
					"creatorId": {
						"type": "string"
					},
					"postId": {
						"type": "string"
					},
					"content": {
						"type": "string"
					}
				},
				"required": [
					"postId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IResPost": {
				"$ref": "#/components/schemas/FlattenMaps_T_"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "workers-server",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "xerexodia ",
			"email": "ibrahimfarhat2018@gmail.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/admin/login": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"201": {
						"description": "welcome"
					}
				},
				"tags": [
					"AdminController"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Pick_IAdmin.email-or-password_"
							}
						}
					}
				}
			}
		},
		"/admin/clients": {
			"get": {
				"operationId": "GetAllClients",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IClient"
									},
									"type": "array"
								}
							}
						}
					},
					"201": {
						"description": "data fetched succesfully"
					}
				},
				"tags": [
					"AdminController"
				],
				"security": [],
				"parameters": []
			}
		},
		"/admin/clients/{id}": {
			"get": {
				"operationId": "GetClient",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IClient"
								}
							}
						}
					},
					"201": {
						"description": "client fetched successfully"
					}
				},
				"tags": [
					"AdminController"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteClient",
				"responses": {
					"201": {
						"description": "client deleted successfully"
					},
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"AdminController"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin/workers": {
			"get": {
				"operationId": "GetAllWorkers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IWorker"
									},
									"type": "array"
								}
							}
						}
					},
					"201": {
						"description": "data fetched succesfully"
					}
				},
				"tags": [
					"AdminController"
				],
				"security": [],
				"parameters": []
			}
		},
		"/admin/workers/{id}": {
			"get": {
				"operationId": "GetWorkerById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IWorker"
								}
							}
						}
					},
					"201": {
						"description": "worker fetched successfully"
					}
				},
				"tags": [
					"AdminController"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteWorker",
				"responses": {
					"201": {
						"description": "worker deleted successfully"
					},
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"AdminController"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin/posts": {
			"get": {
				"operationId": "GetAllPosts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IPost"
									},
									"type": "array"
								}
							}
						}
					},
					"201": {
						"description": "data fetched successfully"
					}
				},
				"tags": [
					"AdminController"
				],
				"security": [],
				"parameters": []
			}
		},
		"/admin/posts/{id}": {
			"get": {
				"operationId": "GetPostById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPost"
								}
							}
						}
					},
					"201": {
						"description": "post fetched successfully"
					}
				},
				"tags": [
					"AdminController"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeletePost",
				"responses": {
					"201": {
						"description": "post deleted successfully"
					},
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"AdminController"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin/materialPosts": {
			"get": {
				"operationId": "GetAllMaterialPosts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IMaterialPost"
									},
									"type": "array"
								}
							}
						}
					},
					"201": {
						"description": "data fetched successfully"
					}
				},
				"tags": [
					"AdminController"
				],
				"security": [],
				"parameters": []
			}
		},
		"/admin/materialPosts/{id}": {
			"get": {
				"operationId": "GetMaterialPostById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IMaterialPost"
								}
							}
						}
					},
					"201": {
						"description": "post fetched successfully"
					}
				},
				"tags": [
					"AdminController"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteMaterialPost",
				"responses": {
					"201": {
						"description": "post deleted successfully"
					},
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"AdminController"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/login": {
			"post": {
				"operationId": "LoginUser",
				"responses": {
					"201": {
						"description": "Welcome back",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IWorker"
										},
										{
											"$ref": "#/components/schemas/IClient"
										},
										{
											"type": "string"
										}
									]
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"msg": {
											"type": "string"
										}
									},
									"required": [
										"msg"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Payload"
							}
						}
					}
				}
			}
		},
		"/login/create": {
			"post": {
				"operationId": "CreateAdmin",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"admin"
				],
				"security": [],
				"parameters": []
			}
		},
		"/register/clients": {
			"post": {
				"operationId": "CreateClient",
				"responses": {
					"201": {
						"description": "Bienvenue",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Pick_IClient.nom-or-prenom-or-password-or-adresse-or-email_"
							}
						}
					}
				}
			}
		},
		"/register/workers": {
			"post": {
				"operationId": "CreateWorker",
				"responses": {
					"201": {
						"description": "Bienvenue",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IWorker"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"msg": {
											"type": "string"
										}
									},
									"required": [
										"msg"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"nom": {
										"type": "string"
									},
									"prenom": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"password": {
										"type": "string"
									},
									"adresse": {
										"type": "string"
									},
									"profession": {
										"type": "string"
									},
									"experience": {
										"type": "string"
									},
									"description": {
										"type": "string"
									},
									"photo": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"nom",
									"prenom",
									"email",
									"password",
									"adresse",
									"profession",
									"experience",
									"description"
								]
							}
						}
					}
				}
			}
		},
		"/comment": {
			"post": {
				"operationId": "AddComment",
				"responses": {
					"201": {
						"description": "created successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IComment"
								}
							}
						}
					}
				},
				"tags": [
					"comments"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IComment"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllComments",
				"responses": {
					"200": {
						"description": "created successfully",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"comments"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "postId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/comment/{id}": {
			"patch": {
				"operationId": "UpdateComment",
				"responses": {
					"201": {
						"description": "updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IComment"
								}
							}
						}
					}
				},
				"tags": [
					"comments"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IComment"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"204": {
						"description": "deleted successfully"
					}
				},
				"tags": [
					"comments"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/materialpost": {
			"post": {
				"operationId": "CreatePost",
				"responses": {
					"201": {
						"description": "created successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IMaterialPost"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"title": {
										"type": "string"
									},
									"adresse": {
										"type": "string"
									},
									"description": {
										"type": "string"
									},
									"workerId": {
										"type": "string"
									},
									"photo": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"title",
									"adresse",
									"description",
									"workerId",
									"photo"
								]
							}
						}
					}
				}
			},
			"get": {
				"operationId": "Getmaterialposts",
				"responses": {
					"201": {
						"description": "fetched successfully",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IMaterialPost"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/materialpost/{materialpostId}": {
			"get": {
				"operationId": "GetPostById",
				"responses": {
					"201": {
						"description": "fetched successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IMaterialPost"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "materialpostId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdatematerialPost",
				"responses": {
					"201": {
						"description": "updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IMaterialPost"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "materialpostId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"title": {
										"type": "string"
									},
									"adresse": {
										"type": "string"
									},
									"description": {
										"type": "string"
									},
									"workerId": {
										"type": "string"
									},
									"photo": {
										"type": "string",
										"format": "binary"
									}
								}
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"201": {
						"description": "deleted successfully"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "materialpostId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/posts": {
			"post": {
				"operationId": "CreatePost",
				"responses": {
					"201": {
						"description": "created successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPost"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"title": {
										"type": "string"
									},
									"adresse": {
										"type": "string"
									},
									"description": {
										"type": "string"
									},
									"clientId": {
										"type": "string"
									},
									"photo": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"title",
									"adresse",
									"description",
									"clientId",
									"photo"
								]
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetPosts",
				"responses": {
					"201": {
						"description": "fetched successfully",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IPost"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			},
			"patch": {
				"operationId": "Reserve",
				"responses": {
					"201": {
						"description": "deleted successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResPost"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"msg": {
											"type": "string"
										}
									},
									"required": [
										"msg"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "postId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "workerId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/posts/client/{clientId}": {
			"get": {
				"operationId": "GetPostsByClientId",
				"responses": {
					"201": {
						"description": "fetched successfully",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IPost"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "clientId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/posts/{postId}": {
			"get": {
				"operationId": "GetPostById",
				"responses": {
					"201": {
						"description": "fetched successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPost"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "postId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdatePost",
				"responses": {
					"201": {
						"description": "updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPost"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "postId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"title": {
										"type": "string"
									},
									"adresse": {
										"type": "string"
									},
									"description": {
										"type": "string"
									},
									"clientId": {
										"type": "string"
									},
									"photo": {
										"type": "string",
										"format": "binary"
									}
								}
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"201": {
						"description": "deleted successfully"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "postId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/clients/update/{id}": {
			"patch": {
				"operationId": "UpdateClient",
				"responses": {
					"201": {
						"description": "Bienvenue",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"clients"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"nom": {
										"type": "string"
									},
									"prenom": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"password": {
										"type": "string"
									},
									"adresse": {
										"type": "string"
									},
									"photo": {
										"type": "string",
										"format": "binary"
									}
								}
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}